Build image with main and Git SHA tags
This commit makes sure that we keep the history of the built images
in quay. It always assigns two tags to the built images.

- main
- Git SHA associated with commit from the PR

This makes sure that main points always to the latest build from the
chatbot repository and the Git SHA tags point to the images built in
the past.