Re-rank search content
This commit adds re-ranking for the results we retrieve from the vector
database. We use the /score API to contact the cross-encoder and to
retrieve a new, more accurate score. The input for the model is
the chunk retrieved from the vector DB and the user's input.

If the chunks from the vector DB are too long, they are chunked even
more to fit the context of the re-rerank model.