Resolve generative model context len exceeded error
Make sure that the full prompt fits the context of the generative model.
The full prompt consists out of:

  - system prompt
  - vector db search results
  - user message

We fill the context with the results from the vector database one by
one. If the context is exceeded, we truncate the last search result
(the one with the lowest similarity score).